GCS: keep browser buttons working regardless of session state
This is a workaround for the fact the modem doens't trigger
a correctly opened state. It also doesn't even start handshaking
by itself, so the session management never even fails.

In general it's useful to keep the browser buttons working for
debugging purposes even if a valid session hasn't been established.