BrushlessGimbal: power up movements
Added a sequence of movements across all channels during
startup that allows verifying the polarity of the outputs
is all correct before tuning. This avoids a frustrating
chicken and the egg scenario between tuning and getting
the signs correct.

The sequence is:
 - ROLL RIGHT
 - ROLL LEFT
 - PITCH UP
 - PITCH DOWN

Note that based on your motor setup it will first snap to
an arbitrary angle before starting.