GCS/Input Config: directly set meta data during calibration
Because the input wizard toggles the metadata a fair number of
times the latency of ack'ing all of them blocks the UI. This
changes the behavior to directly set and store the metadata.
While this might miss one or two settings, it works reliably
for configuring the wizard (and approximates the previous
behavior that has been working for a while).

It also fixes a confusing bug that was occuring at the end
of calibration where the save button was blocking due to
a series of metadata transactions taking up the transaction
queue.