UAVTalk: make relaying work for instances
Previously the UAVTalkRelay method would assume everything
was single instance, and thus did not pass a valid packet
for multiple instance objects. This changes the code to
check if it knows that an object is a multi-instance object.

This requires the modem code be some what in sync with the
flight controller code. However, multiple instance objects
do not change to frequently and this should be tolerable
for now.