RFM22b: if coordinator set connected when received addressed packet
This is not perfect but when the coordinator receives a packet addressed
to it, then set the connected flag to true.

However, there is no code that times out and set that field from true
once the connection is stopped. Essentially, the coordinator has no
expectation to be receiving a stream of data.