GCS/Bind wizard: abstract class for probing boards
This moves the shared functionality of checking for any detected
boards that have a radio into a common class. This class does
not have any UI functionality and thus only works with boards
that are auto connected by USB.

You can also use the main connect/disconnect buttons on the UI
behind the wizard if needed.