Stabilization/MW: rework to use common pid code
This gets the benefit of the bandpass filtered derivative
code and resusing the scaled integral code. These two
changes made it feel a lot more precise and locked in for
me.

Because the Ki term has a strange scaling of input to get
the predicted "rate" we use a setpoint with that strange
value and apply the Kp term outside the pid_apply method.

This also means the derivative term sees that strange
setpoint so currently I am testing with a DerivativeGamma
of 0.