Android: make POI tracking track ground level
The altitude from phones is wildly inaccurate and less useful than
nothing. In addition, the altitude in the UAV is based on the baro
altitude above sea level, and thus should not be compared to the
phone GPS altitude.

Zeroing the POI to the UAV position will not reset the altitude to
always track the height the UAV currently is.