RFM22B: use when packet detected on sync for coordinator connected
Previously the coordinator connected flag was not being set. Now a
packet is sent both directions on the coordinator channel we can
always detect if the pair are connected.

We can also track the fraction of sync pulses received to detected
the link quality more robustly than the current mechanism.