RFM22B: track separate stats on error versus missed packets
We can track easily if synced packets are missed. This is not the
same as receiving a packet and being unable to error correct it.
Now they are separate statistics (although count the same towards
the link statistics).

Note, this doubled the memory required for tracking packet statistics
since we have 5 types.