RFM22B: clean up the connected/disconnected tracking
Connected flag is based on missing a number of consecutive
sync pulses. This means that missing a single sync pulse
will not drop the connected flag. When that happens the
non-coordinator immediately stops hopping and it must wait
to resync. Now it will stay in channel sync through at least
one missed sync pulse.