RFM22B Rcvr: use common code to store to UAVO
Writing the RCVR values to the UAVO is purely for debugging
and not used by any of the other code.

The RFM22BReceiver object is not updated on a timeout
(although any code reading from the driver will be), so
it shows the most recent valid data. Trying to write to
the UAVO from an RTC IRQ seems to sometimes trigger resets,
probably due to a conflict on the UAVO queue.