PiOS DSM: sync code to the tested code
This updates how resolution detection is performed and
makes it safer than previous code (cannot swap after
detecting). This will prevent a situation of being
able to arm in one format and have it change after
that.

All example packets I have found have a channel sequence
starting with 1,5 on the even packet. This adds this check
to the flight code to match the unit tests.