UnitTest/DSM: rewrite resolution detection
The previous code was rather ad hoc and not robust. This
will probe for 10 or 11 bit resolution when the resolution
has not been detected.

Assuming the channel order is steadily increasing is not
valid for DSM streams. Typical orders seen in these packets
are things like:

 Packet 0 -> 1,5,2,3,0,7,6
 Packet 1 -> 1,5,2,3,4,8,9

or for DX7 DSM2

 Packet 0 -> 1,5,2,3
 Packet 1 -> 4,0,7

All the DSM data logs I have managed to collect or find described
online start with Ch1 then Ch5. Use this in the check that it is
either 10 or 11 bit data.