ManualControl: rework arming state machine
The previous code was fairly scattered and hard to verify.
Replaced with a cleaner state machine with documentation
of the funtion.

A functional change of this state machine with the previous
code is that there is a state after the arm and diarm time
that waits for the user to release that position. This
prevents arming and disarming cycling while holding the
corner positions (where the same position is for arming
and disarming).