Transmitter: add option to not timeout arm while autonomous
The default behavior of applying the disarm timeout in all modes
is nice and conservative. However, it is easy to leave the
throttle low while doing navigation and then have it disarm
while flying. This provides an option to avoid that.

Once this has been well validated, eventually we should drop this
option and never disarm while in autonomous modes unless the
user gives the disarm command.