Battery: make sure to load battery value
The module was using an uninitialized variable which
was causing the energy consumed to start at a random
value and the flight time to go to error state.

We could afford to only load this variable once before
the loop, but by doing it each loop it allows GCS to
push changes if desired. This is our more common
practice of making changes in place.