Battery: increase stack size
The previous commit fixed an issue triggering an inappropriate
flight battery time warning. However, it also revealed the stack
usage going extremely low for the Battery module in that case
(6 bytes). That just adds insult to injury if the system is
hitting limits, so this raises the stack size to be safe.