gcs/simulation: Sockets SO_REUSEADDR & TCP_NODELAY
Before, if the simulator crashed, sockets would linger in TIME_WAIT state
and require a timeout before the simulator could be launched again.  This is
now fixed.

Also, request low latency (TCP_NODELAY) service for telemetry data.
Minimizing packets isn't so important compared to getting low telemetry
latency.

This change needs verification in OSX that the simulator still compiles,
and corrects an issue with the previous attempt where a variable was missing
on OSX.