Build: add FLIGHT_BUILD_CONF environment variable
It's a knob, with the values 'default' (which is the default), 'release',
and 'debug'.  In debug, DEBUG is set to YES in the flight firmwares.  In
debug and default, compilation caching is allowed (and ccache is smart
enough to cache default/debug separately because it includes command line
arguments in its hash).  In release, there is no compilation caching (and
jenkins/release build scripts should be updated to use this mode).

This parallels the GCS_BUILD_CONF environment variable which also disables
compilation caching for GCS build.

Altered based on feedback from solidgoldbomb about poor placement of the
DEBUG definition.  It's slightly wordier this way but better.