Sparky: invert telemetry for FrSky Sensor Port
D4R-ii radios (which are quite popular) use the older D8
protocol for telemetry. This is handled in TauLabs by the
UavoFrSKYSensorHubBridge module. Both D8 and D16 (S.Port)
protocols use an inverted serial port. We used the F3
inverted for the S.Port module (UAVOFrskySPortBridge)  but
not the Sensor Hub.

This change allows using telemetry with D4R-ii without an
external inverter.