Build: (partially) deduplicate firmware makefiles
There was a lot of boilerplate that was shared between different
firmware makefiles.  (With very little functional difference, like
conditionals on firmware output formats).  This commit unifies that
boilerplate.

There is still a lot of build system duplication, and I intend to keep
working in this area as a background task.