Uploader: fix safe boot
There was an inconsistency between the bootloader code
and the uploader code. The bootloader swapped the bytes
of the safe word and compared to the constant 0x5AFE.
However, the uploader did not swap those bytes and thus
the safe_word was never valid to trigger a safe boot.

The cleanest thing to do would be remove the swap in the
bootloader, but for ease of the users not updating the
bootloader it swaps in the GCS.