OSD: Make ArtificialHorizon max pitch configurable
Adds a new parameter ArtificialHorizonMaxPitch to the osd page
configuration and utilizes it in the OnScreenDisplay flight module.
The default value of 30 works for slower flights, but various quads
pitch further than 30 degrees forward resulting in an artificial
horizon that is off-screen.  By making it configurable pilots can
adjust it to match their flight profile, or even set different pages
with different maximum pitches depending on how they are flying at
the moment.