Simulator: don't sleep on idle.
Performance seems to fall apart in simulator too much when we sleep on idle.  Maybe it's context switch performance, maybe it's just where I hooked it in.  It needs further study.

Would be really nice to figure this out to lower system resource consumption and to make it clear if the simulator is overtaxed in CPU.  Now, it always uses 100% cpu.