VtolPathFollower: separate path completed logic from inprogress
This more cleanly groups the code that is executed by the path
follower when a path is completed and provides hysteresis to
stop it ever toggling the completed status at the end of a
leg. The code should provide identical output.

The status is only set to completed when the altitude criterion
is achieved. Otherwise we go into endpoint mode. We do _not_ use
the endpoint as the criterion as we are still in a vector flight
leg. However, this could be debated since executing this code
means we already hit the vector criterion. I'm ok either way, but
having both vtol_follower_control_path and vtol_follower_control_simple
set PathStatus is a bad thing..