Sparky2: check in UNTESTED automagically converted kicad files
These files were converted from altium and are NOT entirely
correct. However, they provide a substrate if anyone wants
to create a variant of Sparky2. Otherwise, it is recommended
to use the gerbers, BOM and assembly provided.

The PCB file looks correct.

The SCH file has the major components (surprisingly) but the
power port labels are broken. PR's to fix this by kicad users would
be highly appreciated.