Actuator and PicoC: Remove dependency on UAVO structure order
The Actuator and PicoC modules had a brittle dependency on the auto-
generated code. When the code generating the UAVO changed the order of the
fields in the struct, the code broke badly. This fixes the problem,
although it introduces a minor dependency in the requirement to hardcode
the number of mixers into the modules. This is preferable, because
silent errors can no longer occur. Either it fails on boot, or it fails to
compile, depending on whether new mixers are added or old ones are taken
away.