Matlab: Fix parser to work with timestamps and flash-based logfiles
1) The parser would fail when the UAVO had an optional timestamp.
2) The parser would fail when parsing logfiles created on the onboard
flash. This is because the logfile format changed.

NOTE: Currently, the logic to choose between logfile formats is placed at
the top of the Matlab file. This can lead to a race condition if both
overo and onboardLogger are set to `true`. This will be resolved in a
future commit once the flash logfile identifies itself.