Logging: changed Settings UAVOs to have a minimum period
This ensures that any settings object changes will be captured immediatley.
I foresee a problem with PicoC, since it overwrites settings objects in
order to force real-time control of actuators. This is a separate topic,
but maybe the best thing to do is remove PicoC until it can be rewritten
in a safer manner.