Flight: SPI Speed enhancements for MPU9250_SPI
Enhancement for the SPI clock to speed up data transfer
The data register in MPU can be read up to 20MHz +-10%.
For F4 µC based FC (e.g. Sparky 2) this should result in 21MHz SPI clock, for F1 based µC it should result in 18MHz.

First tests on Quanton with 21MHz were successful.

The speed improvement for Sparky2 should be not as big as for others.
Sparky 2: 10.5MHz -> 21MHz (factor 2)

Signed-off-by: Trex4Git <trex_dev@gmx.net>