UAV Settings: alphabetize UAVObjects
It is very difficult to visually compare two .uav settings files. The saved
settings don't line up because in XML there is no guarantee that the output
file will always have the same node order. This resolves that problem by
having Qt sort the nodes by UAVObject name.

Solution inspired by http://stackoverflow.com/questions/14647638/xslt-sorting-how-to-sort-xml-childnodes-inside-a-parent-node-with-an-attribute