UAVOBrowser: Fix crash when using buttons on non-UAVObjects
When a non-UAVObject is selected and any of the upload/download/save/erase
buttons are used GCS crashes because the returned value is NULL. It is
reasonable for an inexperienced user to click on the buttons without having
first clicked on a UAVObject. Thus this crash seems unnecessary.

This removes the assert and adds an explicit return. There is no adverse
consequence to directly returning, although it might be desirable to alert
the user to the misuse of the buttons, since it is otherwise a silent
failure.