PID Library: Remove unnecessary scaling
The justification for the scaling does not seem to be valid. The original
commit, a09642675cf66650f1cccaf7567878d2f514822b, says "Tweak the integral
calculation so it is scaled in ms internally.  This avoids a loss of
precision on the accumulation.". However, no additional support for the
claim is made in the code or commit, and is contrary to the theory of
floating point precision.

If there is some justification in IEEE-754, it's not immediately obvious.
If dynamic range is a problem, a cleaner solution is to add more bits. It's
dangerous to expect that the all compilers will play nice and make the
multiplication and division by 1000 a worthwhile operation.