UAVSettingsImportExport: Alphabetize attributes in generated XML document
The XML spec intentionally does not care about XML attributes. It considers
them "unimportant". This patch attempts to use XSLT to sort.

Note: According to the original code source,
http://stackoverflow.com/questions/1429991/using-xsl-to-sort-attributes,
this is not guaranteed to work every time. In fact, because XML does not
at all care about attributes, it can be said that this is never guaranteed
to work. However, in practice it works, and it looks likely that if it
works across all implementations of Qt that it will not likely break.

Hopefully, this tides things over until Qt implements a Canonical XML
parser.