Matlab: Add unwrapping of 16-bit UAVO timestamps
The UAVO only holds a 16-bit timestamp, so it will wrap at 65.535s. This
unwraps the time when a subsequent sample is earlier in time than the prior
sample. Note that this can only capture one wrapping period. If there are
multiple wrapping periods hidden inside the data, this will not return an
appropriately unwrapped timestamp.