OpenLRS: prevent beacon being triggered before beacon_period
Previously when failsafe was engaged (1 s after signal loss) a
trigger on the beacon channel could cause the beacon tones to
play. This is convenient if you don't want to wait that 30s (default)
but risks blocking you from reconnecting if there is a false
trigger when in that period.

In practice, this wasn't a significant issue because the beacon
also will not play when armed.