sparky: Enable fastheap
F3 has 8k of CCM-RAM that was not used.

A base Sparky configuration had 7,408 bytes of normal heap free; with
this patch it has 12,972 bytes free. (A 5.5k benefit).

With the resources we've freed up lately, sparky became capable of light
navigation.  With this change, it's pretty much unlimited in
functionality.