Increase stack for battery module.
See #2057 for more details, but after increasing it by 128, I was able
to get it to report 124 bytes free after a bit of jostling (mostly
toying with power levels from a bench supply).

I'm not 100% sure that this is the cause, but I was getting some strange
behavior resulting in watchdog reboots on this particular build.  This
experiment with the battery module indicates it's *possible* that the
battery module was the cause of these issues.

This change increases it by 80 which should be enough to keep things
contained based on current observations.