ManualControl: ignore attempts at arming when not safe
Previously transmitter_control did not account for the state of
the system's safety to arm when checking the sticks. This could
result in the inner module (transmitter_control) considering
itself armed and constantly trying to fire arm events at the
outer module. As soon as the error condition cleared (e.g. GPS
lock occuring) the system would arm and process inputs.