PlatformConfig: Rather large refactoring, targeting only multirotors
In the interests of saving time before full refactoring, this only targets
one platform, i.e. multirotors. The others will follow if this approach is
found to be good.

In a nutshell, this moves away from the GuiConfigDataUnion and toward a
UAVO which holds the platform-specific configuration.

There are many changes which are too numerous to count. However,
overarching changes are:
*) Moving to use the UAVObjectToWidget relationship. This greatly reduces
the number of explicit SIGNAL/SLOT pairs in the code.
*) Singleton to move a signal between two sibling classes.
*) Better abstraction of models.
*) Defining motor names in a more programatic manner.
*) Fewer lines of code.