Actuator: Change return type to be a function of the UAVO
The MixerSettings object variable size can change, so this lets the code
reflect that. It accomplishes this by taking the pointer to the array and
then deferencing it. While this might look weird, taking the pointer to the
array preserves type information so smart compilers can detect accesses
past the end.