Statistics: Functions for efficiently calculating mean and std. dev.
This recognizes that the standard deviation for a data set can be
calculated by the number of elements, a rolling sum of elements, and a
rolling sum of elements squared. The decision is made to use the slightly
less correct std. deviation calculation which divdes by N, instead of N-1.
For large data sets, this makes only a small scalar difference. This
results in a slight gain of computational efficiency.