Sparky2: fix comments labeling current/voltage
The comments indicating which pin was current and voltage
was swapped. The order on the physical header is V/I/DAC
and this is now reflected both in the electrical ordering
(unchanged) and the comments.

This should have no implication for end users as this is
the same order that was present all along.