PIOS FSKDAC: minimum phase between symbols
Because MARK is 1200Hz and SPACE is 2200Hz at 1200bps
there will not be an integer number of cycles for a
space. Thus we have to work out the phase we end each
symbol at and make sure to start the next one at that
phase.

This also drops the code that sends a mid-voltage
between symbols and it constantly outputs MARK when
awaiting a new byte.