Qt Creator Project: Create a Firmware project using Qt Creator.
This includes all files indiscriminately. This is desirable, since this
project file is just a list of all project files. Make must still be
called separately.

Tested and works for debugging and compiling. (Unfortunately, those setups
are on a per-user basis, so by default the project files does *not* work
for compiling and debugging.)