MSPUavoBridge: check in stub
This is a module for the OSD that will systematically make
MSP queries to update the display. This is based off the
opposite end - the UAVOMSPBridge which responds to these
queries from a MWOSD.

At some point, factoring the common MSP parsing code into a
library will make a lot of sense.