Logging: Change FileSectorNum to uint32
For too large files sizes (in the case of the MX25L128 more than 8MB), the
logging download would fail because FileSectorNum was limited to 16 bits
(65535 sectors). This increases the maximum number of sectors, although
FileSectorNum is currently copied into an int32, so at most this gives 2^31
sectors. Still, this is far greater than any likely memory requirement in
the short term.