VTOLPathFollower: Fix incorrect - sign
The assumption was made that a forward accel would have a negative value.
This is confusing and unnecessary. This commit places the negative sign
inside the calculation of the pitch angle, which indeed must be negative
for forward acceleration.

In addition, aligned some of the symmetric code so that the differences
stand out more clearly.