MS5xxx baro: Use different math for different sensors
The math for each sensor in the MS5xxx line is slightly different. So far,
it is only a difference in the scalar value of the multipliers and
divisors. This likely means that the section can be further refactored by
only handling the differences in the scalars separately.

Also moved the conversion math into a separate function. This function is
free of any PiOS calls and so can be checked with unit tests. This will be
helpful for refactoring any of the math