Qt Creator: Introduce script for automatic user project file configuration
Currently only configures flight project file. It creates a build target
for each board, as well as the POSIX simulator.

Assumes that the user does not want to use multiple cores for a build. This
is an unfortunate default, since most users will be able to use multiple
cores, but it seems the safest way to build on a generic platform.