Matlab/LQG: basic LQG logistics
This sets up an LQG controller using a Kalman filter
to estimate instantaneous torque and rotation rate and
an LQR controller on this to optimally achieve the
desired rate.

In principle this should work better than a PID controller
on gyros as it can benefit from smoothness assumptions on
the flight behavior and thus be less sensitive to derivative
noise. Of course, it should respond well during a collision
so ultimately trust the gyros fairly tightly.

For fixed system properties and tuning parameters, the LQR
calculation should not change so could be initially done
entirely on the desktop side (or of course, at first in
Matlab).