Matlab/LQG: continuous implementation of model
This model is in continuous time and uses the current value of
the parmameters to calculate the step response. It is worth
noting that LQI controller in matlab is subtly different than
the one used in the code. Specifically, the instantaneous error
is not used for the state error but is only passed to the
integral.