LQG/LQR Solver: update for per-axis model
This now independently solves the problem for roll
pitch and yaw which should reduce memory and runtime.

It uses the same dynamical model for angle and rate
controllers as it is the difference between having
an integral that you try and hold at zero versus
while controlling rate versus controlling angle
directly.