LQG/KF: convert to a fixed gain filter (model is time invariant)
This introduces a number of changes. The most dramatic is that
it uses fixed gains for feedback to compute the state estimate
which removes the requirement to track the covariance, as well
as a number of calculations.

 x_t+1 = A x_t + B u_t + L(y - C*x)

L is calculated from the discrete algebraic riccati equation and
either can be offline or online (currently implemented but not
integrated).

In addition, underlying it uses a symbolically discretized version
of the continuous time state model - so all the underlying math
and convergence properties can be done in the continuous domain.