Constants: Query KPA2HECTAPASCAL
This codebase looks a bit stale, but I was looking through it and noticed this constant is incorrectly defined.  I haven't fixed it, as it is likely used in the code somewhere and may have a hack to fix it (I haven't looked).  Up to you what you do with it.